requirements.txt
SpeechRecognition>=3.10.0
pyttsx3>=2.90
requests>=2.28.0
flask>=2.2.0
nltk>=3.8
PyQt5>=5.15.6
python-dotenv>=1.0.0
Werkzeug>=2.2.0